HTTP_API/requirements.txt
azure-functions
textblob